+++ lib/shamsi_date.dart
@@ -6,6 +6,7 @@ library shamsi_date;
 
 export 'package:shamsi_date/src/date.dart';
 export 'package:shamsi_date/src/date_formatter.dart';
+export 'package:shamsi_date/src/exceptions.dart';
 export 'package:shamsi_date/src/gregorian/gregorian_date.dart';
 export 'package:shamsi_date/src/gregorian/gregorian_formatter.dart';
 export 'package:shamsi_date/src/jalali/jalali_date.dart';